pyproject.toml
[tool.poetry]
name = "profile_stack"
version = "1.0.0"
description = "🚀 Display your tech stack on your GitHub profile's README"
authors = ["Matthew Gleich <matthewgleich@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
loguru = "^0.5.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
